Agregadas vistas para listado y creación de libros

+++ database/migrations/2020_11_04_211548_create_libros_table.php
@@ -18,7 +18,7 @@ class CreateLibrosTable extends Migration
             $table->string('nombre', 100);
             $table->string('autor', 255)->default('anonimo');
             $table->string('editorial', 100);
-            $table->unsignedTinyInteger('edicion');
+            $table->unsignedSmallInteger('edicion');
             $table->year('anio');
             $table->unsignedSmallInteger('paginas');
             $table->timestamps();
